refactor(react-apollo-hooks): Move main modules to separate files.
