+++ src/utils.ts
@@ -1,4 +1,5 @@
 import ApolloClient from 'apollo-client';
+import isPlainObject from 'lodash/isPlainObject';
 
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
@@ -11,4 +12,18 @@ export function assertApolloClient(client: null | ApolloClient<any>) {
         'ApolloClient instance in via props.'
     );
   }
+}
+
+export default function objToKey(obj: Record<string, any>): null | string {
+  if (!obj) {
+    return null;
+  }
+  const sortedObj = Object.keys(obj)
+    .sort()
+    .reduce((result: Record<string, any>, key) => {
+      const value = obj[key];
+      result[key] = isPlainObject(value) ? objToKey(obj[key]) : obj[key];
+      return result;
+    }, {});
+  return JSON.stringify(sortedObj);
 }
